feat: make wasm-bindgen-test a dev-dependency, not a regular dependency
