+++ bindings_wasm/Cargo.toml
@@ -15,7 +15,6 @@ wasm-bindgen = "0.2"
 web-sys = { version = "0.3.61", features = ["console", "Window", "Storage"] }
 rand = { version = "0.7.3", features = ["wasm-bindgen"] }
 getrandom = { version = "0.2", features = ["js"] }
-wasm-bindgen-test = "0.3"
 
 [dev-dependencies]
 wasm-bindgen-test = "0.3"
