+++ bindings_wasm/run_tests.sh
