+++ bindings_wasm/src/lib.rs
@@ -3,7 +3,6 @@ mod local_storage_persistence;
 use local_storage_persistence::LocalStoragePersistence;
 use std::sync::Mutex;
 use wasm_bindgen::prelude::*;
-use wasm_bindgen_test::*;
 use xmtp::{Client, ClientBuilder};
 
 static CLIENT_LIST: Mutex<Vec<Client<LocalStoragePersistence>>> = Mutex::new(Vec::new());
@@ -41,19 +40,4 @@ pub fn client_read_from_persistence(
     let mut clients = CLIENT_LIST.lock().unwrap();
     let client = clients.get_mut(client_id).expect("Client not found");
     client.read_from_persistence(key)
-}
-
-wasm_bindgen_test_configure!(run_in_browser);
-#[wasm_bindgen_test]
-fn can_pass_persistence_methods() {
-    let client_id = client_create();
-    assert_eq!(
-        client_read_from_persistence(client_id, "foo".to_string()).unwrap(),
-        None
-    );
-    client_write_to_persistence(client_id, "foo".to_string(), b"bar").unwrap();
-    assert_eq!(
-        client_read_from_persistence(client_id, "foo".to_string()).unwrap(),
-        Some(b"bar".to_vec())
-    );
 }
